Separate out the remaining f32/i7/i8 implementations into their own files
