+++ libs/simdvec/native/src/vec/c/amd64/vec_i8_1.cpp
@@ -22,196 +22,6 @@
 #define STRIDE_BYTES_LEN sizeof(__m256i) // Must be a power of 2
 #endif
 
-// Accumulates acc += dot(pa, pb) for unsigned 7-bit int lanes (32 bytes per step).
-template<int offsetRegs>
-inline void fmai7u(__m256i& acc, const int8_t* pa, const int8_t* pb) {
-    constexpr int lanes = offsetRegs * STRIDE_BYTES_LEN;
-    const __m256i a = _mm256_loadu_si256((const __m256i_u*)(pa + lanes));
-    const __m256i b = _mm256_loadu_si256((const __m256i_u*)(pb + lanes));
-    const __m256i vab = _mm256_maddubs_epi16(a, b);
-    const __m256i ones = _mm256_set1_epi16(1);
-    acc = _mm256_add_epi32(_mm256_madd_epi16(ones, vab), acc);
-}
-
-static inline int32_t doti7u_inner(const int8_t* a, const int8_t* b, const int32_t dims) {
-    int32_t res = 0;
-    int i = 0;
-    if (dims >= STRIDE_BYTES_LEN) {
-        i = dims & ~(STRIDE_BYTES_LEN - 1);
-
-        constexpr int batches = 4;
-        static_assert(batches % 2 == 0, "batches must be even");
-        constexpr int half_batches = batches / 2;
-        constexpr int batch_stride = batches * STRIDE_BYTES_LEN;
-        constexpr int half_batch_stride = half_batches * STRIDE_BYTES_LEN;
-        const int8_t* pa = a;
-        const int8_t* pb = b;
-
-        __m256i acc[batches];
-        apply_indexed<batches>([&](auto I) {
-            acc[I] = _mm256_setzero_si256();
-        });
-
-        const int8_t* pa_end = a + (i & ~(batch_stride - 1));
-        while (pa < pa_end) {
-            apply_indexed<batches>([&](auto I) {
-                fmai7u<I>(acc[I], pa, pb);
-            });
-            pa += batch_stride;
-            pb += batch_stride;
-        }
-
-        pa_end = a + (i & ~(half_batch_stride - 1));
-        while (pa < pa_end) {
-            apply_indexed<half_batches>([&](auto I) {
-                fmai7u<I>(acc[I], pa, pb);
-            });
-            pa += half_batch_stride;
-            pb += half_batch_stride;
-        }
-
-        pa_end = a + i;
-        while (pa < pa_end) {
-            fmai7u<0>(acc[0], pa, pb);
-            pa += STRIDE_BYTES_LEN;
-            pb += STRIDE_BYTES_LEN;
-        }
-
-        __m256i total_sum = tree_reduce<batches, __m256i, _mm256_add_epi32>(acc);
-        res = mm256_reduce_epi32<_mm_add_epi32>(total_sum);
-    }
-    // scalar tail
-    for (; i < dims; i++) {
-        res += dot_scalar(a[i], b[i]);
-    }
-    return res;
-}
-
-EXPORT int32_t vec_doti7u(const int8_t* a, const int8_t* b, const int32_t dims) {
-    return doti7u_inner(a, b, dims);
-}
-
-EXPORT void vec_doti7u_bulk(const int8_t* a, const int8_t* b, const int32_t dims, const int32_t count, f32_t* results) {
-    call_i8_bulk<int8_t, sequential_mapper, doti7u_inner>(a, b, dims, dims, NULL, count, results);
-}
-
-EXPORT void vec_doti7u_bulk_offsets(
-    const int8_t* a,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results) {
-    call_i8_bulk<int8_t, offsets_mapper, doti7u_inner>(a, b, dims, pitch, offsets, count, results);
-}
-
-EXPORT void vec_doti7u_bulk_sparse(
-    const void* const* addresses,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t count,
-    f32_t* results) {
-    call_i8_bulk<const int8_t*, sparse_mapper, doti7u_inner>((const int8_t* const*)addresses, b, dims, 0, NULL, count, results);
-}
-
-// Accumulates acc += sqr_distance(pa, pb) for unsigned 7-bit int lanes (32 bytes per step).
-template<int offsetRegs>
-inline void sqri7u(__m256i& acc, const int8_t* pa, const int8_t* pb) {
-    constexpr int lanes = offsetRegs * STRIDE_BYTES_LEN;
-    const __m256i a = _mm256_loadu_si256((const __m256i_u*)(pa + lanes));
-    const __m256i b = _mm256_loadu_si256((const __m256i_u*)(pb + lanes));
-    const __m256i dist = _mm256_sub_epi8(a, b);
-    const __m256i abs_dist = _mm256_sign_epi8(dist, dist);
-    const __m256i sqr = _mm256_maddubs_epi16(abs_dist, abs_dist);
-    const __m256i ones = _mm256_set1_epi16(1);
-    acc = _mm256_add_epi32(_mm256_madd_epi16(ones, sqr), acc);
-}
-
-static inline int32_t sqri7u_inner(const int8_t* a, const int8_t* b, const int32_t dims) {
-    int32_t res = 0;
-    int i = 0;
-    if (dims >= STRIDE_BYTES_LEN) {
-        i = dims & ~(STRIDE_BYTES_LEN - 1);
-
-        constexpr int batches = 4;
-        static_assert(batches % 2 == 0, "batches must be even");
-        constexpr int half_batches = batches / 2;
-        constexpr int batch_stride = batches * STRIDE_BYTES_LEN;
-        constexpr int half_batch_stride = half_batches * STRIDE_BYTES_LEN;
-        const int8_t* pa = a;
-        const int8_t* pb = b;
-
-        __m256i acc[batches];
-        apply_indexed<batches>([&](auto I) {
-            acc[I] = _mm256_setzero_si256();
-        });
-
-        const int8_t* pa_end = a + (i & ~(batch_stride - 1));
-        while (pa < pa_end) {
-            apply_indexed<batches>([&](auto I) {
-                sqri7u<I>(acc[I], pa, pb);
-            });
-            pa += batch_stride;
-            pb += batch_stride;
-        }
-
-        pa_end = a + (i & ~(half_batch_stride - 1));
-        while (pa < pa_end) {
-            apply_indexed<half_batches>([&](auto I) {
-                sqri7u<I>(acc[I], pa, pb);
-            });
-            pa += half_batch_stride;
-            pb += half_batch_stride;
-        }
-
-        pa_end = a + i;
-        while (pa < pa_end) {
-            sqri7u<0>(acc[0], pa, pb);
-            pa += STRIDE_BYTES_LEN;
-            pb += STRIDE_BYTES_LEN;
-        }
-
-        __m256i total_sum = tree_reduce<batches, __m256i, _mm256_add_epi32>(acc);
-        res = mm256_reduce_epi32<_mm_add_epi32>(total_sum);
-    }
-    // scalar tail
-    for (; i < dims; i++) {
-        res += sqr_scalar(a[i], b[i]);
-    }
-    return res;
-}
-
-EXPORT int32_t vec_sqri7u(const int8_t* a, const int8_t* b, const int32_t dims) {
-    return sqri7u_inner(a, b, dims);
-}
-
-EXPORT void vec_sqri7u_bulk(const int8_t* a, const int8_t* b, const int32_t dims, const int32_t count, f32_t* results) {
-    call_i8_bulk<int8_t, sequential_mapper, sqri7u_inner>(a, b, dims, dims, NULL, count, results);
-}
-
-EXPORT void vec_sqri7u_bulk_offsets(
-    const int8_t* a,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results) {
-    call_i8_bulk<int8_t, offsets_mapper, sqri7u_inner>(a, b, dims, pitch, offsets, count, results);
-}
-
-EXPORT void vec_sqri7u_bulk_sparse(
-    const void* const* addresses,
-    const int8_t* b,
-    const int32_t dims,
-    const int32_t count,
-    f32_t* results) {
-    call_i8_bulk<const int8_t*, sparse_mapper, sqri7u_inner>((const int8_t* const*)addresses, b, dims, 0, NULL, count, results);
-}
-
-// --- byte vectors
-
 /*
  * AVX2 doesn't have sign-symmetric 8-bit operations,
  * so we have to sign-extend to 16-bits and operate on those values
@@ -632,179 +442,4 @@ EXPORT void vec_sqri8_bulk_sparse(
         count,
         results
     );
-}
-
-// --- single precision floats
-
-/*
- * Float bulk operation. Iterates over 4 sequential vectors at a time.
- *
- * Template parameters:
- * mapper: gets the nth vector from the input array.
- * inner_op: SIMD per-dimension vector operation, takes a, b, sum, returns new sum
- * scalar_op: scalar per-dimension vector operation, takes a, b, returns sum
- * bulk_tail: complete vector comparison on a single vector
- *
- * This should compile to a single inline method, with no function callouts.
- */
-template <
-    typename TData,
-    const f32_t*(*mapper)(const TData*, const int32_t, const int32_t*, const int32_t),
-    __m256(*inner_op)(const __m256, const __m256, const __m256),
-    f32_t(*scalar_op)(const f32_t, const f32_t),
-    f32_t(*bulk_tail)(const f32_t*, const f32_t*, const int32_t),
-    int batches = 4
->
-static inline void call_f32_bulk(
-    const TData* a,
-    const f32_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results
-) {
-    int c = 0;
-
-    for (; c + batches - 1 < count; c += batches) {
-        // Pointers to the current batch of input vectors, resolved via mapper.
-        // as[0] points to the vector for index 0, [1] for index 1, etc
-        const f32_t* as[batches];
-        __m256 sums[batches] = {};
-        apply_indexed<batches>([&](auto I) {
-            as[I] = mapper(a, c + I, offsets, pitch);
-            sums[I] = _mm256_setzero_ps();
-        });
-
-        int32_t i = 0;
-        // do <batches> vectors at a time, iterating through the dimensions in parallel
-        // Each __m256 holds 8 floats
-        constexpr int stride = sizeof(__m256) / sizeof(f32_t);
-        for (; i < (dims & ~(stride - 1)); i += stride) {
-            __m256 bi = _mm256_loadu_ps(b + i);
-            apply_indexed<batches>([&](auto I) {
-                sums[I] = inner_op(_mm256_loadu_ps(as[I] + i), bi, sums[I]);
-            });
-        }
-
-        f32_t res[batches];
-        apply_indexed<batches>([&](auto I) {
-            res[I] = mm256_reduce_ps<_mm_add_ps>(sums[I]);
-        });
-
-        // dimensions tail
-        for (; i < dims; i++) {
-            apply_indexed<batches>([&](auto I) {
-                res[I] += scalar_op(as[I][i], b[i]);
-            });
-        }
-
-        // this should be turned into direct value copies by the compiler
-        std::copy_n(res, batches, results + c);
-    }
-
-    // Tail-handling: remaining vectors
-    for (; c < count; c++) {
-        const f32_t* a0 = mapper(a, c, offsets, pitch);
-        results[c] = bulk_tail(a0, b, dims);
-    }
-}
-
-// const f32_t* a  pointer to the first float vector
-// const f32_t* b  pointer to the second float vector
-// const int32_t elementCount  the number of floating point elements
-EXPORT f32_t vec_dotf32(const f32_t* a, const f32_t* b, const int32_t elementCount) {
-    constexpr int batches = 4;
-
-    __m256 sums[batches];
-    apply_indexed<batches>([&](auto I) {
-        sums[I] = _mm256_setzero_ps();
-    });
-
-    int i = 0;
-    // each value has <elements> floats, and we iterate over <stride> floats at a time
-    constexpr int elements = sizeof(__m256) / sizeof(f32_t);
-    constexpr int stride = sizeof(__m256) / sizeof(f32_t) * batches;
-    for (; i < (elementCount & ~(stride - 1)); i += stride) {
-        apply_indexed<batches>([&](auto I) {
-            sums[I] = _mm256_fmadd_ps(_mm256_loadu_ps(a + i + I * elements), _mm256_loadu_ps(b + i + I * elements), sums[I]);
-        });
-    }
-
-    // Combine all partial sums
-    __m256 total_sum = tree_reduce<batches, __m256, _mm256_add_ps>(sums);
-    f32_t result = mm256_reduce_ps<_mm_add_ps>(total_sum);
-
-    for (; i < elementCount; ++i) {
-        result += dot_scalar(a[i], b[i]);
-    }
-
-    return result;
-}
-
-EXPORT void vec_dotf32_bulk(const f32_t* a, const f32_t* b, const int32_t dims, const int32_t count, f32_t* results) {
-    call_f32_bulk<f32_t, sequential_mapper, _mm256_fmadd_ps, dot_scalar<f32_t>, vec_dotf32>(a, b, dims, dims, NULL, count, results);
-}
-
-EXPORT void vec_dotf32_bulk_offsets(
-    const f32_t* a,
-    const f32_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results) {
-    call_f32_bulk<f32_t, offsets_mapper, _mm256_fmadd_ps, dot_scalar<f32_t>, vec_dotf32>(a, b, dims, pitch / sizeof(f32_t), offsets, count, results);
-}
-
-static inline __m256 sqrf32_vector(__m256 a, __m256 b, __m256 sum) {
-    __m256 diff = _mm256_sub_ps(a, b);
-    return _mm256_fmadd_ps(diff, diff, sum);
-}
-
-// const f32_t* a  pointer to the first float vector
-// const f32_t* b  pointer to the second float vector
-// const int32_t elementCount  the number of floating point elements
-EXPORT f32_t vec_sqrf32(const f32_t* a, const f32_t* b, const int32_t elementCount) {
-    constexpr int batches = 4;
-
-    __m256 sums[batches];
-    apply_indexed<batches>([&](auto I) {
-        sums[I] = _mm256_setzero_ps();
-    });
-
-    int i = 0;
-    // each value has <elements> floats, and we iterate over <stride> floats at a time
-    constexpr int elements = sizeof(__m256) / sizeof(f32_t);
-    constexpr int stride = sizeof(__m256) / sizeof(f32_t) * batches;
-    for (; i < (elementCount & ~(stride - 1)); i += stride) {
-        apply_indexed<batches>([&](auto I) {
-            sums[I] = sqrf32_vector(_mm256_loadu_ps(a + i + I * elements), _mm256_loadu_ps(b + i + I * elements), sums[I]);
-        });
-    }
-
-    // reduce all partial sums
-    __m256 total_sum = tree_reduce<batches, __m256, _mm256_add_ps>(sums);
-    f32_t result = mm256_reduce_ps<_mm_add_ps>(total_sum);
-
-    for (; i < elementCount; ++i) {
-        result += sqr_scalar(a[i], b[i]);
-    }
-
-    return result;
-}
-
-EXPORT void vec_sqrf32_bulk(const f32_t* a, const f32_t* b, const int32_t dims, const int32_t count, f32_t* results) {
-    call_f32_bulk<f32_t, sequential_mapper, sqrf32_vector, sqr_scalar<f32_t>, vec_sqrf32>(a, b, dims, dims, NULL, count, results);
-}
-
-EXPORT void vec_sqrf32_bulk_offsets(
-    const f32_t* a,
-    const f32_t* b,
-    const int32_t dims,
-    const int32_t pitch,
-    const int32_t* offsets,
-    const int32_t count,
-    f32_t* results) {
-    call_f32_bulk<f32_t, offsets_mapper, sqrf32_vector, sqr_scalar<f32_t>, vec_sqrf32>(a, b, dims, pitch / sizeof(f32_t), offsets, count, results);
 }
